+++ docs/user_guide/changelog.rst
@@ -15,6 +15,7 @@ Bug Fixes
 ---------
 
 - Fixed BatchNorm running statistics, now using the unbiased estimator to update ``running_var`` at training time.
+- Fixed scalar tensor inputs to the graph not working
 
 v2.6 (Poplar SDK 2.6)
 =====================